requirements.txt
attr==0.3.2
attrs==21.2.0
bitarray==2.3.6
black==21.9b0
click==8.0.1
iniconfig==1.1.1
mypy==0.910
mypy-extensions==0.4.3
numpy==1.21.2
packaging==21.0
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
regex==2021.9.24
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
Pillow==9.3.0
zstandard==0.18.0
